Skip changelog parsing on autocut PRs
Skips the changelog parsing process for autocut PRs as they will
fail most of the time.

For example, here:
https://github.com/opensearch-project/OpenSearch-Dashboards/pull/8195

The workflow will showed it failed which could cause maintainers to lax
on reviewing the PR when glancing at it since it seems like it has a failure.
Even though it was just a backport PR.

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>